pyproject.toml
[project]
name = "dragonbot-new"
version = "0.1.0"
description = "A discord bot with main functionality being a ticket manager."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "bs4>=0.0.2",
    "discord-py>=2.5.2",
    "numpy>=2.2.6",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.4",
]
